[Fix] 코드 리뷰 수정
