+++ src/screens/Mypage/HelpScreen.tsx
@@ -7,7 +7,7 @@ import { UndoIcon } from "@/assets/icons";
 import { helpScreenStyles as styles } from "./HelpScreen.styles";
 import { HELP_DATA } from "./MypageData";
 import type { NativeStackScreenProps } from "@react-navigation/native-stack";
-import { RootStackParamList } from "@/screens/Test/TestPlace";
+import { RootStackParamList } from "@/types/navigation";
 
 type Props = NativeStackScreenProps<RootStackParamList, "Help">;
